requirements.txt
flask
pymysql
requests
bs4
lxml
random
numpy
werkzeug.security